server/requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dill==0.3.8
filelock==3.15.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
fsspec==2024.6.0
gcloud==0.18.3
googleapis-common-protos==1.63.1
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python-headless
oauth2client==4.1.3
opencv-python==4.10.0.82
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.27.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyngrok==7.1.6
pyparsing==3.1.2
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.12.1
tbb==2021.12.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.31
ultralytics-thop==0.2.8
urllib3==1.26.18
Werkzeug==3.0.3